Replace and with plus, and add and_every
Before you could say `.every(Monday).and(Wednesday)`, and would not behave like you'd expect it to
as a reader of English. Now it's `.every(Monday).plus(Wednesday)` which doesn't make sense.

However, the expected behaviour of `.every(Monday).and(Wednesday)` was appealing, so I added
`.and_every` which adds an additional schedule for the job.