Add custom timezone support
The default is chrono::Local, which matches current behaviour and won't
break existing code, but with the `with_tz` constructor, you can use
the timezone of your choice.

This updates the code, and changes the documentation to reflect the
behaviour of `Scheduler::new`, but doesn't
update the Readme or examples.